fix: folders

+++ cognite/powerops/resync/core/main.py
@@ -17,10 +17,7 @@ from cognite.client import CogniteClient
 from cognite.client.data_classes.data_modeling import DataModelId, MappedProperty, ViewList
 from cognite.client.exceptions import CogniteAPIError
 from cognite_toolkit.cdf import Common, build, deploy  # type: ignore[import-untyped]
-from cognite_toolkit.cdf_tk.utils import (  # type: ignore[import-untyped]
-    CDFToolConfig,
-    calculate_directory_hash,
-)
+from cognite_toolkit.cdf_tk.utils import CDFToolConfig, calculate_directory_hash  # type: ignore[import-untyped]
 from rich import print
 from rich.panel import Panel
 from typer import Context
@@ -96,11 +93,24 @@ def init(client: PowerOpsClient | None, is_dev: bool = False, dry_run: bool = Fa
 
     build_folder = Path.cwd() / "build"
     # Bug in toolkit that places the containers and views in the wrong folder
-    for file in itertools.chain((build_folder / "containers").glob("*.yaml"), (build_folder / "views").glob("*.yaml")):
+    folders = [
+        "containers",
+        "views",
+        "assets",
+        "bid_configuration",
+        "bid_document",
+        "bid_matrix",
+        "bid_method",
+        "function_inputs",
+        "function_outputs",
+        "scenario",
+    ]
+    for file in itertools.chain(*((build_folder / folder).glob("*.yaml") for folder in folders)):
         shutil.move(file, build_folder / "data_models" / file.name)
 
-    for folder in ["containers", "views"]:
-        shutil.rmtree(build_folder / folder)
+    for folder in folders:
+        if (build_folder / folder).exists():
+            shutil.rmtree(build_folder / folder)
 
     loader = DataModelLoader(build_folder)
     schema = loader.load()
